update on make_phase_split
