+++ idaes/models_extra/column_models/condenser.py
@@ -42,7 +42,6 @@ from idaes.core import (
     ControlVolume0DBlock,
     declare_process_block_class,
     EnergyBalanceType,
-    MomentumBalanceType,
     MaterialBalanceType,
     UnitModelBlockData,
     useDefault,
@@ -56,6 +55,7 @@ from idaes.core.util.exceptions import (
     InitializationError,
 )
 from idaes.core.solvers import get_solver
+from idaes.models_extra.column_models.column_util import make_phase_split
 
 _log = idaeslog.getLogger(__name__)
 
@@ -170,7 +170,6 @@ see property package for documentation.}""",
 
     def build(self):
         """Build the model.
-
         Args:
             None
         Returns:
@@ -244,9 +243,28 @@ see property package for documentation.}""",
 
         self._make_ports()
 
+        def rule_reflux(self):
+            return (
+                self.reflux_ratio * (1 + self.reflux_ratio)**-1
+            )
+
+        self.expr1 = Expression(rule=rule_reflux)
+
         if self.config.condenser_type == CondenserType.totalCondenser:
 
-            self._make_splits_total_condenser()
+            make_phase_split(self.control_volume,
+                    port=self.reflux,
+                    phase=self._liquid_set,
+                    side_sf=(self.expr1),
+                    equipmentType=CondenserType.totalCondenser
+                )
+
+            make_phase_split(self.control_volume,
+                    port=self.distillate,
+                    phase=self._liquid_set,
+                    side_sf=(1- self.expr1),
+                    equipmentType=CondenserType.totalCondenser
+                )
 
             if self.config.temperature_spec == TemperatureSpec.atBubblePoint:
                 # Option 1: if true, condition for total condenser
@@ -280,7 +298,26 @@ see property package for documentation.}""",
                 )
 
         else:
-            self._make_splits_partial_condenser()
+
+            make_phase_split(self.control_volume,
+                    port=self.reflux,
+                    phase=self._liquid_set,
+                    side_sf=(self.expr1),
+                    equipmentType=CondenserType.partialCondenser
+                )
+
+            make_phase_split(self.control_volume,
+                    port=self.distillate,
+                    phase=self._liquid_set,
+                    side_sf=(1- self.expr1),
+                    equipmentType=CondenserType.partialCondenser
+                )
+            make_phase_split(self.control_volume,
+                    port=self.vapor_outlet,
+                    phase=self._vapor_set,
+                    side_sf=1,
+                    equipmentType=CondenserType.partialCondenser
+                )
 
         # Add object reference to variables of the control volume
         # Reference to the heat duty
@@ -309,512 +346,7 @@ see property package for documentation.}""",
                 noruleinit=True, doc="Vapor outlet port from a " "partial condenser"
             )
         # Add codnenser specific variables
-        self.reflux_ratio = Var(initialize=1, doc="Reflux ratio for the condenser")
-
-    def _make_splits_total_condenser(self):
-
-        # Get dict of Port members and names
-        member_list = self.control_volume.properties_out[0].define_port_members()
-
-        # Create references and populate the reflux, distillate ports
-        for k in member_list:
-
-            local_name = member_list[k].local_name
-            # Create references and populate the intensive variables
-            if "flow" not in local_name:
-                if not member_list[k].is_indexed():
-                    var = self.control_volume.properties_out[:].component(local_name)
-                else:
-                    var = self.control_volume.properties_out[:].component(local_name)[
-                        ...
-                    ]
-
-                # add the reference and variable name to the reflux port
-                self.reflux.add(Reference(var), k)
-
-                # add the reference and variable name to the distillate port
-                self.distillate.add(Reference(var), k)
-
-            elif "flow" in local_name:
-                # Create references and populate the extensive variables
-                # This is for vars that are not indexed
-                if not member_list[k].is_indexed():
-                    # Expression for reflux flow and relation to the
-                    # reflux_ratio variable
-
-                    def rule_reflux_flow(self, t):
-                        return self.control_volume.properties_out[t].component(
-                            local_name
-                        ) * (self.reflux_ratio / (1 + self.reflux_ratio))
-
-                    self.e_reflux_flow = Expression(
-                        self.flowsheet().time, rule=rule_reflux_flow
-                    )
-                    self.reflux.add(self.e_reflux_flow, k)
-
-                    # Expression for distillate flow and relation to the
-                    # reflux_ratio variable
-                    def rule_distillate_flow(self, t):
-                        return self.control_volume.properties_out[t].component(
-                            local_name
-                        ) / (1 + self.reflux_ratio)
-
-                    self.e_distillate_flow = Expression(
-                        self.flowsheet().time, rule=rule_distillate_flow
-                    )
-                    self.distillate.add(self.e_distillate_flow, k)
-                else:
-                    # Create references and populate the extensive variables
-                    # This is for vars that are indexed by phase, comp or both.
-                    index_set = member_list[k].index_set()
-
-                    def rule_reflux_flow(self, t, *args):
-                        return self.control_volume.properties_out[t].component(
-                            local_name
-                        )[args] * (self.reflux_ratio / (1 + self.reflux_ratio))
-
-                    self.e_reflux_flow = Expression(
-                        self.flowsheet().time, index_set, rule=rule_reflux_flow
-                    )
-                    self.reflux.add(self.e_reflux_flow, k)
-
-                    def rule_distillate_flow(self, t, *args):
-                        return self.control_volume.properties_out[t].component(
-                            local_name
-                        )[args] / (1 + self.reflux_ratio)
-
-                    self.e_distillate_flow = Expression(
-                        self.flowsheet().time, index_set, rule=rule_distillate_flow
-                    )
-                    self.distillate.add(self.e_distillate_flow, k)
-            else:
-                raise PropertyNotSupportedError(
-                    "Unrecognized names for flow variables encountered while "
-                    "building the condenser ports."
-                )
-
-    def _make_splits_partial_condenser(self):
-        # Get dict of Port members and names
-        member_list = self.control_volume.properties_out[0].define_port_members()
-
-        # Create references and populate the reflux, distillate ports
-        for k in member_list:
-
-            local_name = member_list[k].local_name
-
-            # Create references and populate the intensive variables
-            if (
-                "flow" not in local_name
-                and "frac" not in local_name
-                and "enth" not in local_name
-            ):
-                if not member_list[k].is_indexed():
-                    var = self.control_volume.properties_out[:].component(local_name)
-                else:
-                    var = self.control_volume.properties_out[:].component(local_name)[
-                        ...
-                    ]
-
-                # add the reference and variable name to the reflux port
-                self.reflux.add(Reference(var), k)
-
-                # add the reference and variable name to the distillate port
-                self.distillate.add(Reference(var), k)
-
-                # add the reference and variable name to the
-                # vapor outlet port
-                self.vapor_outlet.add(Reference(var), k)
-
-            elif "frac" in local_name:
-
-                # Mole/mass frac is typically indexed
-                index_set = member_list[k].index_set()
-
-                # if state var is not mole/mass frac by phase
-                if "phase" not in local_name:
-                    if "mole" in local_name:  # check mole basis/mass basis
-
-                        # The following conditionals are required when a
-                        # mole frac or mass frac is a state var i.e. will be
-                        # a port member. This gets a bit tricky when handling
-                        # non-conventional systems when you have more than one
-                        # liquid or vapor phase. Hence, the logic here is that
-                        # the mole frac that should be present in the liquid or
-                        # vapor port should be computed by accounting for
-                        # multiple liquid or vapor phases if present. For the
-                        # classical VLE system, this holds too.
-                        if hasattr(
-                            self.control_volume.properties_out[0],
-                            "mole_frac_phase_comp",
-                        ) and hasattr(
-                            self.control_volume.properties_out[0], "flow_mol_phase"
-                        ):
-                            flow_phase_comp = False
-                            local_name_frac = "mole_frac_phase_comp"
-                            local_name_flow = "flow_mol_phase"
-                        elif hasattr(
-                            self.control_volume.properties_out[0], "flow_mol_phase_comp"
-                        ):
-                            flow_phase_comp = True
-                            local_name_flow = "flow_mol_phase_comp"
-                        else:
-                            raise PropertyNotSupportedError(
-                                "No mole_frac_phase_comp or flow_mol_phase or"
-                                " flow_mol_phase_comp variables encountered "
-                                "while building ports for the condenser. "
-                            )
-                    elif "mass" in local_name:
-                        if hasattr(
-                            self.control_volume.properties_out[0],
-                            "mass_frac_phase_comp",
-                        ) and hasattr(
-                            self.control_volume.properties_out[0], "flow_mass_phase"
-                        ):
-                            flow_phase_comp = False
-                            local_name_frac = "mass_frac_phase_comp"
-                            local_name_flow = "flow_mass_phase"
-                        elif hasattr(
-                            self.control_volume.properties_out[0],
-                            "flow_mass_phase_comp",
-                        ):
-                            flow_phase_comp = True
-                            local_name_flow = "flow_mass_phase_comp"
-                        else:
-                            raise PropertyNotSupportedError(
-                                "No mass_frac_phase_comp or flow_mass_phase or"
-                                " flow_mass_phase_comp variables encountered "
-                                "while building ports for the condenser."
-                            )
-                    else:
-                        raise PropertyNotSupportedError(
-                            "No mass frac or mole frac variables encountered "
-                            " while building ports for the condenser. "
-                            "phase_frac as a state variable is not "
-                            "supported with distillation unit models."
-                        )
-
-                    # Rule for liquid phase mole fraction
-                    def rule_liq_frac(self, t, i):
-                        if not flow_phase_comp:
-                            sum_flow_comp = sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_frac
-                                )[p, i]
-                                * self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p]
-                                for p in self._liquid_set
-                            )
-
-                            return sum_flow_comp / sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p]
-                                for p in self._liquid_set
-                            )
-                        else:
-                            sum_flow_comp = sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p, i]
-                                for p in self._liquid_set
-                            )
-
-                            return sum_flow_comp / sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p, i]
-                                for p in self._liquid_set
-                                for i in self.config.property_package.component_list
-                            )
-
-                    self.e_liq_frac = Expression(
-                        self.flowsheet().time, index_set, rule=rule_liq_frac
-                    )
-
-                    # Rule for vapor phase mass/mole fraction
-                    def rule_vap_frac(self, t, i):
-                        if not flow_phase_comp:
-                            sum_flow_comp = sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_frac
-                                )[p, i]
-                                * self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p]
-                                for p in self._vapor_set
-                            )
-                            return sum_flow_comp / sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p]
-                                for p in self._vapor_set
-                            )
-                        else:
-                            sum_flow_comp = sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p, i]
-                                for p in self._vapor_set
-                            )
-
-                            return sum_flow_comp / sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p, i]
-                                for p in self._vapor_set
-                                for i in self.config.property_package.component_list
-                            )
-
-                    self.e_vap_frac = Expression(
-                        self.flowsheet().time, index_set, rule=rule_vap_frac
-                    )
-
-                    # add the reference and variable name to the reflux port
-                    self.reflux.add(self.e_liq_frac, k)
-
-                    # add the reference and variable name to the
-                    # distillate port
-                    self.distillate.add(self.e_liq_frac, k)
-
-                    # add the reference and variable name to the
-                    # vapor port
-                    self.vapor_outlet.add(self.e_vap_frac, k)
-                else:
-
-                    # Assumes mole_frac_phase or mass_frac_phase exist as
-                    # state vars in the port and therefore access directly
-                    # from the state block.
-                    var = self.control_volume.properties_out[:].component(local_name)[
-                        ...
-                    ]
-
-                    # add the reference and variable name to the reflux port
-                    self.reflux.add(Reference(var), k)
-
-                    # add the reference and variable name to the distillate port
-                    self.distillate.add(Reference(var), k)
-            elif "flow" in local_name:
-                if "phase" not in local_name:
-
-                    # Assumes that here the var is total flow or component
-                    # flow. However, need to extract the flow by phase from
-                    # the state block. Expects to find the var
-                    # flow_mol_phase or flow_mass_phase in the state block.
-
-                    # Check if it is not indexed by component list and this
-                    # is total flow
-                    if not member_list[k].is_indexed():
-                        # if state var is not flow_mol/flow_mass by phase
-                        local_name_flow = local_name + "_phase"
-
-                        # Rule for vap phase flow
-                        def rule_vap_flow(self, t):
-                            return sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p]
-                                for p in self._vapor_set
-                            )
-
-                        self.e_vap_flow = Expression(
-                            self.flowsheet().time, rule=rule_vap_flow
-                        )
-
-                        # Rule to link the liq phase flow to the reflux
-                        def rule_reflux_flow(self, t):
-                            return sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p]
-                                for p in self._liquid_set
-                            ) * (self.reflux_ratio / (1 + self.reflux_ratio))
-
-                        self.e_reflux_flow = Expression(
-                            self.flowsheet().time, rule=rule_reflux_flow
-                        )
-
-                        # Rule to link the liq flow to the distillate
-                        def rule_distillate_flow(self, t):
-                            return sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p]
-                                for p in self._liquid_set
-                            ) / (1 + self.reflux_ratio)
-
-                        self.e_distillate_flow = Expression(
-                            self.flowsheet().time, rule=rule_distillate_flow
-                        )
-
-                    else:
-                        # when it is flow comp indexed by component list
-                        str_split = local_name.split("_")
-                        if len(str_split) == 3 and str_split[-1] == "comp":
-                            local_name_flow = (
-                                str_split[0] + "_" + str_split[1] + "_phase_" + "comp"
-                            )
-
-                        # Get the indexing set i.e. component list
-                        index_set = member_list[k].index_set()
-
-                        # Rule for vap phase flow to the vapor outlet
-                        def rule_vap_flow(self, t, i):
-                            return sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p, i]
-                                for p in self._vapor_set
-                            )
-
-                        self.e_vap_flow = Expression(
-                            self.flowsheet().time, index_set, rule=rule_vap_flow
-                        )
-
-                        # Rule to link the liq flow to the reflux
-                        def rule_reflux_flow(self, t, i):
-                            return sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p, i]
-                                for p in self._liquid_set
-                            ) * (self.reflux_ratio / (1 + self.reflux_ratio))
-
-                        self.e_reflux_flow = Expression(
-                            self.flowsheet().time, index_set, rule=rule_reflux_flow
-                        )
-
-                        # Rule to link the liq flow to the distillate
-                        def rule_distillate_flow(self, t, i):
-                            return sum(
-                                self.control_volume.properties_out[t].component(
-                                    local_name_flow
-                                )[p, i]
-                                for p in self._liquid_set
-                            ) / (1 + self.reflux_ratio)
-
-                        self.e_distillate_flow = Expression(
-                            self.flowsheet().time, index_set, rule=rule_distillate_flow
-                        )
-
-                    # add the reference and variable name to the reflux port
-                    self.reflux.add(self.e_reflux_flow, k)
-
-                    # add the reference and variable name to the
-                    # distillate port
-                    self.distillate.add(self.e_distillate_flow, k)
-
-                    # add the reference and variable name to the
-                    # distillate port
-                    self.vapor_outlet.add(self.e_vap_flow, k)
-            elif "enth" in local_name:
-                if "phase" not in local_name:
-                    # assumes total mixture enthalpy (enth_mol or enth_mass)
-                    # and hence should not be indexed by phase
-                    if not member_list[k].is_indexed():
-                        # if state var is not enth_mol/enth_mass
-                        # by phase, add _phase string to extract the right
-                        # value from the state block
-                        local_name_enth = local_name + "_phase"
-                    else:
-                        raise PropertyPackageError(
-                            "Enthalpy is indexed but the variable "
-                            "name does not reflect the presence of an index. "
-                            "Please follow the naming convention outlined "
-                            "in the documentation for state variables."
-                        )
-
-                    # NOTE:pass phase index when generating expression only
-                    # when multiple liquid or vapor phases detected
-                    # else ensure consistency with state vars and do not
-                    # add phase index to the port members. Hence, the check
-                    # for length of local liq and vap phase sets.
-
-                    # Rule for vap enthalpy. Setting the enthalpy to the
-                    # enth_mol_phase['Vap'] value from the state block
-                    def rule_vap_enth(self, t):
-                        return sum(
-                            self.control_volume.properties_out[t].component(
-                                local_name_enth
-                            )[p]
-                            for p in self._vapor_set
-                        )
-
-                    self.e_vap_enth = Expression(
-                        self.flowsheet().time, rule=rule_vap_enth
-                    )
-
-                    # Rule to link the liq enthalpy to the reflux.
-                    # Setting the enthalpy to the
-                    # enth_mol_phase['Liq'] value from the state block
-                    def rule_reflux_enth(self, t):
-                        return sum(
-                            self.control_volume.properties_out[t].component(
-                                local_name_enth
-                            )[p]
-                            for p in self._liquid_set
-                        )
-
-                    self.e_reflux_enth = Expression(
-                        self.flowsheet().time, rule=rule_reflux_enth
-                    )
-
-                    # Rule to link the liq flow to the distillate.
-                    # Setting the enthalpy to the
-                    # enth_mol_phase['Liq'] value from the state block
-                    def rule_distillate_enth(self, t):
-                        return sum(
-                            self.control_volume.properties_out[t].component(
-                                local_name_enth
-                            )[p]
-                            for p in self._liquid_set
-                        )
-
-                    self.e_distillate_enth = Expression(
-                        self.flowsheet().time, rule=rule_distillate_enth
-                    )
-
-                    # add the reference and variable name to the reflux port
-                    self.reflux.add(self.e_reflux_enth, k)
-
-                    # add the reference and variable name to the
-                    # distillate port
-                    self.distillate.add(self.e_distillate_enth, k)
-
-                    # add the reference and variable name to the
-                    # distillate port
-                    self.vapor_outlet.add(self.e_vap_enth, k)
-                elif "phase" in local_name:
-                    # assumes enth_mol_phase or enth_mass_phase.
-                    # This is an intensive property, you create a direct
-                    # reference irrespective of the reflux, distillate and
-                    # vap_outlet
-
-                    # Rule for vap flow
-                    if not member_list[k].is_indexed():
-                        var = self.control_volume.properties_out[:].component(
-                            local_name
-                        )
-                    else:
-                        var = self.control_volume.properties_out[:].component(
-                            local_name
-                        )[...]
-
-                    # add the reference and variable name to the reflux port
-                    self.reflux.add(Reference(var), k)
-
-                    # add the reference and variable name to the distillate port
-                    self.distillate.add(Reference(var), k)
-
-                    # add the reference and variable name to the
-                    # vapor outlet port
-                    self.vapor_outlet.add(Reference(var), k)
-                else:
-                    raise PropertyNotSupportedError(
-                        "Unrecognized enthalpy state variable encountered "
-                        "while building ports for the condenser. Only total "
-                        "mixture enthalpy or enthalpy by phase are supported."
-                    )
+        self.reflux_ratio = Var(initialize=0.5, doc="Reflux ratio for the condenser")
 
     def initialize(
         self, state_args=None, solver=None, optarg=None, outlvl=idaeslog.NOTSET
