+++ idaes/models_extra/column_models/tray_column.py
@@ -39,6 +39,8 @@ from idaes.core.util.exceptions import ConfigurationError, InitializationError
 from idaes.core.util.config import is_physical_parameter_block
 from idaes.core.solvers import get_solver
 
+from idaes.core.util.initialization import propagate_state
+
 _log = idaeslog.getLogger(__name__)
 
 
@@ -206,7 +208,6 @@ see property package for documentation.}""",
 
     def build(self):
         """Build the model.
-
         Args:
             None
         Returns:
@@ -396,20 +397,6 @@ see property package for documentation.}""",
             destination=self.stripping_section[self.config.number_of_trays].vap_in,
         )
 
-    def propagate_stream_state(self, source=None, destination=None):
-        """
-        This method is used during initialization to propage state values
-        between any two ports.
-
-        Args:
-            source : source port
-            destination : destination port
-        """
-        for v in source.vars:
-            for i in destination.vars[v]:
-                if not destination.vars[v][i].fixed:
-                    destination.vars[v][i].value = value(source.vars[v][i])
-
     def initialize(
         self,
         state_args_feed=None,
@@ -431,27 +418,29 @@ see property package for documentation.}""",
             solver=solver, optarg=optarg, outlvl=outlvl
         )
 
-        self.propagate_stream_state(
-            source=self.feed_tray.vap_out, destination=self.condenser.inlet
+        propagate_state(
+            destination=self.condenser.inlet,
+            source=self.feed_tray.vap_out
         )
 
         self.condenser.initialize(solver=solver, optarg=optarg, outlvl=outlvl)
 
-        self.propagate_stream_state(
-            source=self.feed_tray.liq_out, destination=self.reboiler.inlet
+        propagate_state(
+            destination=self.reboiler.inlet,
+            source=self.feed_tray.liq_out
         )
 
         self.reboiler.initialize(solver=solver, optarg=optarg, outlvl=outlvl)
 
         # initialize the rectification section
         for i in self._rectification_index:
-            self.propagate_stream_state(
-                source=self.condenser.reflux,
+            propagate_state(
                 destination=self.rectification_section[i].liq_in,
+                source=self.condenser.reflux
             )
-            self.propagate_stream_state(
-                source=self.feed_tray.vap_out,
+            propagate_state(
                 destination=self.rectification_section[i].vap_in,
+                source=self.feed_tray.vap_out
             )
             if i == 1:
                 rect_liq_flags = self.rectification_section[i].initialize(
@@ -476,13 +465,13 @@ see property package for documentation.}""",
 
         # initialize the stripping section
         for i in self._stripping_index:
-            self.propagate_stream_state(
-                source=self.feed_tray.liq_out,
+            propagate_state(
                 destination=self.stripping_section[i].liq_in,
+                source=self.feed_tray.liq_out
             )
-            self.propagate_stream_state(
-                source=self.reboiler.vapor_reboil,
+            propagate_state(
                 destination=self.stripping_section[i].vap_in,
+                source=self.reboiler.vapor_reboil
             )
             if i == self.config.feed_tray_location + 1:
                 strip_liq_flags = self.stripping_section[i].initialize(
